fix: refactor geofence repository structure and update attendance config service to handle geofencing disable logic
