--- backend/src/main/java/com/skapp/community/common/repository/WorkLocationGeofenceRepository.java
@@ -1,9 +1,0 @@
-package com.skapp.community.common.repository;
-
-public interface WorkLocationGeofenceRepository {
-
-	void deleteAllGeofences();
-
-	void clearAddressesForGeofencedLocations();
-
-}
