+++ backend/src/main/java/com/skapp/community/common/repository/WorkLocationRepository.java
@@ -13,4 +13,6 @@ public interface WorkLocationRepository {
 
 	List<WorkLocation> findAllWorkLocationsOrderByNameAsc();
 
+	void clearAddressesForGeofencedLocations();
+
 }
