+++ backend/src/main/java/com/skapp/community/common/repository/impl/WorkLocationRepositoryImpl.java
@@ -2,19 +2,24 @@ package com.skapp.community.common.repository.impl;
 
 import com.skapp.community.common.model.WorkLocation;
 import com.skapp.community.common.model.WorkLocation_;
+import com.skapp.community.common.model.WorkLocationGeofence;
+import com.skapp.community.common.model.WorkLocationGeofence_;
 import com.skapp.community.common.payload.request.WorkLocationFilterDto;
 import com.skapp.community.common.repository.WorkLocationRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.CriteriaUpdate;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
+import jakarta.persistence.criteria.Subquery;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -86,6 +91,23 @@ public class WorkLocationRepositoryImpl implements WorkLocationRepository {
 		}
 
 		return entityManager.createQuery(countQuery).getSingleResult();
+	}
+
+	@Override
+	@Transactional
+	public void clearAddressesForGeofencedLocations() {
+		CriteriaBuilder cb = entityManager.getCriteriaBuilder();
+		CriteriaUpdate<WorkLocation> update = cb.createCriteriaUpdate(WorkLocation.class);
+		Root<WorkLocation> root = update.from(WorkLocation.class);
+
+		Subquery<Long> subquery = update.subquery(Long.class);
+		Root<WorkLocationGeofence> geofenceRoot = subquery.from(WorkLocationGeofence.class);
+		subquery.select(geofenceRoot.get(WorkLocationGeofence_.workLocation).get(WorkLocation_.workLocationId));
+
+		update.set(root.get(WorkLocation_.address), (String) null);
+		update.where(root.get(WorkLocation_.workLocationId).in(subquery));
+
+		entityManager.createQuery(update).executeUpdate();
 	}
 
 }
